Bug 565660 workbench restriction executor
 Move all the work to internal.
