+++ bundles/org.eclipse.passage.lic.jface/META-INF/MANIFEST.MF
@@ -19,7 +19,7 @@ Export-Package: org.eclipse.passage.lic.internal.jface.dialogs;x-internal:=true,
  org.eclipse.passage.lic.jface,
  org.eclipse.passage.lic.jface.actions,
  org.eclipse.passage.lic.jface.dialogs,
- org.eclipse.passage.lic.jface.dialogs.licensing;x-internal:=true,
+ org.eclipse.passage.lic.internal.jface.dialogs.licensing;x-internal:=true,
  org.eclipse.passage.lic.jface.resource,
  org.eclipse.passage.lic.jface.viewers,
  org.eclipse.passage.lic.jface.widgets
