+++ bundles/org.eclipse.passage.lic.jface/src/org/eclipse/passage/lic/internal/jface/dialogs/licensing/WorkbenchNotification.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.jface.dialogs.licensing;
+package org.eclipse.passage.lic.internal.jface.dialogs.licensing;
 
 import java.util.function.Supplier;
